+++ .eslintrc.js
@@ -160,7 +160,7 @@ module.exports = {
       },
     },
     {
-      files: ['packages/react-flight-dom-webpack/**/*.js'],
+      files: ['packages/react-transport-dom-webpack/**/*.js'],
       globals: {
         __webpack_chunk_load__: true,
         __webpack_require__: true,
