+++ fixtures/flight-browser/index.html
@@ -19,8 +19,8 @@
     <script src="../../build/node_modules/react/umd/react.development.js"></script>
     <script src="../../build/node_modules/react-dom/umd/react-dom.development.js"></script>
     <script src="../../build/node_modules/react-dom/umd/react-dom-server.browser.development.js"></script>
-    <script src="../../build/node_modules/react-flight-dom-webpack/umd/react-flight-dom-webpack-server.browser.development.js"></script>
-    <script src="../../build/node_modules/react-flight-dom-webpack/umd/react-flight-dom-webpack.development.js"></script>
+    <script src="../../build/node_modules/react-transport-dom-webpack/umd/react-transport-dom-webpack-server.browser.development.js"></script>
+    <script src="../../build/node_modules/react-transport-dom-webpack/umd/react-transport-dom-webpack.development.js"></script>
     <script src="https://unpkg.com/babel-standalone@6/babel.js"></script>
     <script type="text/babel">
       let Suspense = React.Suspense;
