+++ fixtures/flight/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import ReactFlightDOMClient from 'react-flight-dom-webpack';
+import ReactFlightDOMClient from 'react-transport-dom-webpack';
 import App from './App';
 
 let data = ReactFlightDOMClient.createFromFetch(fetch('http://localhost:3001'));
