+++ packages/react-client/src/forks/ReactFlightClientHostConfig.dom-browser.js
@@ -9,4 +9,4 @@
 
 export * from 'react-client/src/ReactFlightClientHostConfigBrowser';
 export * from 'react-client/src/ReactFlightClientHostConfigStream';
-export * from 'react-flight-dom-webpack/src/ReactFlightClientWebpackBundlerConfig';
+export * from 'react-transport-dom-webpack/src/ReactFlightClientWebpackBundlerConfig';
