+++ packages/react-transport-dom-relay/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "react-flight-dom-relay",
+  "name": "react-transport-dom-relay",
   "version": "0.1.0",
   "private": true,
   "repository": {
     "type" : "git",
     "url" : "https://github.com/facebook/react.git",
-    "directory": "packages/react-flight-dom-relay"
+    "directory": "packages/react-transport-dom-relay"
   },
   "dependencies": {
     "object-assign": "^4.1.1",
