+++ packages/react-transport-dom-webpack/npm/plugin.js
@@ -1,0 +1,7 @@
+'use strict';
+
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./cjs/react-transport-dom-webpack-plugin.production.min.js');
+} else {
+  module.exports = require('./cjs/react-transport-dom-webpack-plugin.development.js');
+}
