+++ packages/react-transport-dom-webpack/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "react-flight-dom-webpack",
-  "description": "React Flight bindings for DOM using Webpack.",
+  "name": "react-transport-dom-webpack",
+  "description": "React Transport bindings for DOM using Webpack. This is intended to be integrated into meta-frameworks. It is not intended to be imported directly.",
   "version": "0.1.0",
   "private": true,
   "keywords": [
@@ -28,7 +28,7 @@
   "repository": {
     "type" : "git",
     "url" : "https://github.com/facebook/react.git",
-    "directory": "packages/react-flight-dom-webpack"
+    "directory": "packages/react-transport-dom-webpack"
   },
   "engines": {
     "node": ">=0.10.0"
