+++ packages/react-transport-dom-webpack/plugin.js
