+++ packages/react-transport-dom-webpack/src/ReactFlightClientWebpackBundlerConfig.js
