+++ packages/react-transport-dom-webpack/src/ReactFlightDOMServerBrowser.js
