+++ scripts/rollup/bundles.js
@@ -241,21 +241,21 @@ const bundles = [
   {
     bundleTypes: [NODE_DEV, NODE_PROD, UMD_DEV, UMD_PROD],
     moduleType: RENDERER,
-    entry: 'react-flight-dom-webpack/server.browser',
+    entry: 'react-transport-dom-webpack/server.browser',
     global: 'ReactFlightDOMServer',
     externals: ['react', 'react-dom/server'],
   },
   {
     bundleTypes: [NODE_DEV, NODE_PROD],
     moduleType: RENDERER,
-    entry: 'react-flight-dom-webpack/server.node',
+    entry: 'react-transport-dom-webpack/server.node',
     global: 'ReactFlightDOMServer',
     externals: ['react', 'react-dom/server'],
   },
   {
     bundleTypes: [NODE_DEV, NODE_PROD],
     moduleType: RENDERER,
-    entry: 'react-flight-dom-webpack/server-runtime',
+    entry: 'react-transport-dom-webpack/server-runtime',
     global: 'ReactFlightDOMServerRuntime',
     externals: ['react'],
   },
@@ -264,7 +264,7 @@ const bundles = [
   {
     bundleTypes: [NODE_DEV, NODE_PROD, UMD_DEV, UMD_PROD],
     moduleType: RENDERER,
-    entry: 'react-flight-dom-webpack',
+    entry: 'react-transport-dom-webpack',
     global: 'ReactFlightDOMClient',
     externals: ['react'],
   },
@@ -273,7 +273,7 @@ const bundles = [
   {
     bundleTypes: [NODE_DEV, NODE_PROD],
     moduleType: RENDERER_UTILS,
-    entry: 'react-flight-dom-webpack/plugin',
+    entry: 'react-transport-dom-webpack/plugin',
     global: 'ReactFlightWebpackPlugin',
     externals: [],
     babel: opts =>
@@ -293,7 +293,7 @@ const bundles = [
   {
     bundleTypes: [FB_WWW_DEV, FB_WWW_PROD],
     moduleType: RENDERER,
-    entry: 'react-flight-dom-relay/server',
+    entry: 'react-transport-dom-relay/server',
     global: 'ReactFlightDOMRelayServer',
     externals: [
       'react',
@@ -304,7 +304,7 @@ const bundles = [
   {
     bundleTypes: [FB_WWW_DEV, FB_WWW_PROD],
     moduleType: RENDERER,
-    entry: 'react-flight-dom-relay/server-runtime',
+    entry: 'react-transport-dom-relay/server-runtime',
     global: 'ReactFlightDOMRelayServerRuntime',
     externals: ['react', 'ReactFlightDOMRelayServerIntegration'],
   },
@@ -313,7 +313,7 @@ const bundles = [
   {
     bundleTypes: [FB_WWW_DEV, FB_WWW_PROD],
     moduleType: RENDERER,
-    entry: 'react-flight-dom-relay',
+    entry: 'react-transport-dom-relay',
     global: 'ReactFlightDOMRelayClient',
     externals: ['react', 'ReactFlightDOMRelayClientIntegration'],
   },
