+++ scripts/shared/inlinedHostConfigs.js
@@ -13,20 +13,20 @@ module.exports = [
       'react-dom',
       'react-dom/testing',
       'react-dom/unstable-fizz.node',
-      'react-flight-dom-webpack/server.node',
-      'react-flight-dom-webpack/server-runtime',
-      'react-flight-dom-webpack',
+      'react-transport-dom-webpack/server.node',
+      'react-transport-dom-webpack/server-runtime',
+      'react-transport-dom-webpack',
     ],
     paths: [
       'react-dom',
       'react-dom/unstable-fizz',
       'react-dom/unstable-fizz.node',
       'react-dom/src/server/ReactDOMFizzServerNode.js', // react-dom/unstable-fizz.node
-      'react-flight-dom-webpack',
-      'react-flight-dom-webpack/server',
-      'react-flight-dom-webpack/server.node',
-      'react-flight-dom-webpack/server-runtime',
-      'react-flight-dom-webpack/src/ReactFlightDOMServerNode.js', // react-flight-dom-webpack/server.browser
+      'react-transport-dom-webpack',
+      'react-transport-dom-webpack/server',
+      'react-transport-dom-webpack/server.node',
+      'react-transport-dom-webpack/server-runtime',
+      'react-transport-dom-webpack/src/ReactFlightDOMServerNode.js', // react-transport-dom-webpack/server.browser
       'react-client/src/ReactFlightClientStream.js', // We can only type check this in streaming configurations.
       'react-interactions',
     ],
@@ -39,19 +39,19 @@ module.exports = [
       'react-dom',
       'react-dom/testing',
       'react-dom/unstable-fizz.browser',
-      'react-flight-dom-webpack/server.browser',
-      'react-flight-dom-webpack/server-runtime',
-      'react-flight-dom-webpack',
+      'react-transport-dom-webpack/server.browser',
+      'react-transport-dom-webpack/server-runtime',
+      'react-transport-dom-webpack',
     ],
     paths: [
       'react-dom',
       'react-dom/testing',
       'react-dom/unstable-fizz.browser',
       'react-dom/src/server/ReactDOMFizzServerBrowser.js', // react-dom/unstable-fizz.browser
-      'react-flight-dom-webpack',
-      'react-flight-dom-webpack/server.browser',
-      'react-flight-dom-webpack/server-runtime',
-      'react-flight-dom-webpack/src/ReactFlightDOMServerBrowser.js', // react-flight-dom-webpack/server.browser
+      'react-transport-dom-webpack',
+      'react-transport-dom-webpack/server.browser',
+      'react-transport-dom-webpack/server-runtime',
+      'react-transport-dom-webpack/src/ReactFlightDOMServerBrowser.js', // react-transport-dom-webpack/server.browser
       'react-client/src/ReactFlightClientStream.js', // We can only type check this in streaming configurations.
     ],
     isFlowTyped: true,
@@ -88,11 +88,11 @@ module.exports = [
   {
     shortName: 'dom-relay',
     entryPoints: [
-      'react-flight-dom-relay',
-      'react-flight-dom-relay/server',
-      'react-flight-dom-relay/server-runtime',
+      'react-transport-dom-relay',
+      'react-transport-dom-relay/server',
+      'react-transport-dom-relay/server-runtime',
     ],
-    paths: ['react-dom', 'react-flight-dom-relay'],
+    paths: ['react-dom', 'react-transport-dom-relay'],
     isFlowTyped: true,
     isServerSupported: true,
   },
